+++ src/Core/View/CalcInfo.xaml.cs
@@ -15,8 +15,29 @@ using System.Windows.Shapes;
 using System.Globalization;
 // Markup is so ContentProperty works correctly
 using System.Windows.Markup;
+using System.Windows.Interactivity;
 
 namespace NinjaCalc.Core.View {
+
+    // Used on sub-controls of an expander to bubble the mouse wheel scroll event up 
+    public sealed class BubbleScrollEvent : Behavior<UIElement> {
+        protected override void OnAttached() {
+            base.OnAttached();
+            AssociatedObject.PreviewMouseWheel += AssociatedObject_PreviewMouseWheel;
+        }
+
+        protected override void OnDetaching() {
+            AssociatedObject.PreviewMouseWheel -= AssociatedObject_PreviewMouseWheel;
+            base.OnDetaching();
+        }
+
+        void AssociatedObject_PreviewMouseWheel(object sender, MouseWheelEventArgs e) {
+            e.Handled = true;
+            var e2 = new MouseWheelEventArgs(e.MouseDevice, e.Timestamp, e.Delta);
+            e2.RoutedEvent = UIElement.MouseWheelEvent;
+            AssociatedObject.RaiseEvent(e2);
+        }
+    }
 
 
     /// <summary>