+++ changelog.md
@@ -6,6 +6,7 @@ v0.3.7
 - Made the application startup in maximised state (full screen size), closes #53.
 - Made a reusable "Info" WPF user control, closes #52.
 - Added note to README about how installing new version should automatically overwrite old version, closes #54.
+- Fixed bug where calculator view didn't scroll when mouse was over expanded "Info" text, closes #55.
 
 v0.3.6
 ------